Moved unicode stuff into a separate file
