+++ icons.py
@@ -1,19 +1,4 @@
-import sys
-
-# Support Unicode literals with both Python 2 and 3
-if sys.version < '3':
-    import codecs
-    def u(x):
-        return codecs.unicode_escape_decode(x)[0]
-
-    def uchr(x):
-        return unichr(x)
-else:
-    def u(x):
-        return x
-
-    def uchr(x):
-        return chr(x)
+from uchr import u, uchr
 
 icons = {
     "adjust": u("\uf042"),
